refactor(utils): consolidate and centralize imports via utils/index
