+++ src/app/receive/select-token.tsx
@@ -1,6 +1,5 @@
 import { assetConfig } from '@/config/assets';
-import getDisplaySymbol from '@/utils/get-display-symbol';
-import { getRecentTokens, addToRecentTokens } from '@/utils/recent-tokens';
+import { getRecentTokens, addToRecentTokens, getDisplaySymbol } from '@/utils';
 import { useWallet } from '@tetherto/wdk-react-native-provider';
 import { useDebouncedNavigation } from '@/hooks/use-debounced-navigation';
 import { ArrowLeft, Search, X } from 'lucide-react-native';
