+++ src/components/TokenDetails.tsx
@@ -1,13 +1,10 @@
 import { networkConfigs } from '@/config/networks';
-import formatAmount from '@/utils/format-amount';
-import formatTokenAmount from '@/utils/format-token-amount';
-import formatUSDValue from '@/utils/format-usd-value';
 import { AssetTicker, NetworkType } from '@tetherto/wdk-react-native-provider';
 import { Send } from 'lucide-react-native';
 import React from 'react';
 import { Alert, Image, ScrollView, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import { colors } from '@/constants/colors';
-import { gt } from '@/utils/bignumber';
+import { formatAmount, formatTokenAmount, formatUSDValue, gt } from '@/utils';
 
 interface TokenNetworkBalance {
   network: string;
